Allow custom SSL certificates
Some libraries have SSL certificates which are not recognized trusted by Android (or, sometimes only as of Android 3.x/4.x). We do want to use them anyways. Don't want to allow all certificates or turn off hostname validation, just specific ones we ship in our BKS keystore. This is not less secure than what we have now, because we for now trust not more than two certificates more than before -- and those two were verified (and can only be used for this specific library, due to the domain name they are limited to).

fixing #156 and #91