SISIS: fix problem when more than 30 items are lent (#519)
When more than 30 items are lent (or, presumably, reserved) in a SISIS OPAC, the link to the fourth page of items (10 items per page) is not visible from the first page. This changes the implementation so that each page is checked for links to previously unseen pages instead of only going to all links visible on the first page.