Move branch field down if status text is too long
The branch field is displayed on top of the right half of the status field.
If the status text is longer than half of the total width and thus would
overlap the branch field, move the branch field below the status field
but keep it right aligned to the center.
The branch field's visibility is initially set to GONE as it's not used
in the reservations adapter.