SLUB: Test for getResultById
This test is not ideal as the test should just verify that parseResultById
will be called with correct id but not actually call it. So it should be

Mockito.doReturn(DetailedItem()).`when`(slub).parseResultById(Matchers.any(), Matchers.any())
...
verify(slub).parseResultById("id/123", JSONObject(response))

but this throws an IllegalStateException: Mockito.any() must not be null because of
the not nullable Kotlin function signature of parseResultById. None of the solutions
from  https://stackoverflow.com/questions/30305217/is-it-possible-to-use-mockito-in-kotlin
work. See also https://github.com/mockito/mockito/issues/1255.
The only possible solution seems to be switching to MockK.