Make links in copy views clickable
Needed for SLUB where copy status may be empty and a link is provided
to the institute/chair where this book might be available.
(prior to Jan 2021 the link was provided in the status field e.g. as
"Verfügbarkeit bitte in <a href="...">Prof ...</a> erfragen")
See also https://stackoverflow.com/a/20647011/3944322.