Refresh data after prolongAll or prolongMultiple without results list
prolongAll and prolongMultiple return a ProlongAllResult object that may or may not
contain a list of the individual success values. Some APIs (e.g. Koha or SLUB) don't
populate this results list (ProlongAllResult.results == null) and only provide a
general message (MultistepResult.message). In this case the account information was
not updated after a successful renewal. This PR forces data update after a successful
renewal even if the results list is null.